lmao

+++ zooma-api/Repositories/IOrderRepository.cs
@@ -8,7 +8,7 @@ namespace Repositories
     {
         int CreateOrder(short customerId, List<CartItemDTO> cartItems);
         Order GetOrdersById(int orderId);
-        List<Order> GetOrdersByCustomerId(int customerID);
+        List<Order> GetOrdersByCustomerId(short customerID);
         List<OrderDetail> GetOrderDetailsByOrderId(int orderId);
 
         void updateRefundOrder(int orderId);
